library fixes +event dispatching example
